Remove sax dependency from emfio
Module 'sax' was one of the dependencies of 'emfio'. This is also
mentioned in:

   https://wiki.documentfoundation.org/Development/Build_System

Previously, module 'emfio' was directly depending on 'drawinglayer'
and 'sax'. But the build is fine when 'sax' is removed from
'emfio/Library_emfio.mk'. Also there is no mention of sax when
invoking 'git grep sax' inside emfio folder.

With this patch, the 'emfio' dependency on 'sax' is removed.

Change-Id: Ib6f9dc3df33865b6fc912d02b27b01232a6a3fb8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123988
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>