ScriptForge - (SF_Calc) new Region() property
The Region() property expects one mandatory argument,
a range given as a string.

It returns a new range as a string, including the sheet name,
corresponding with the smallest area
- containing the input range
- completely surrounded with empty cells

It is complementary to Offset() and A1Style() to make easy the
definition of new ranges relative to known ranges

The property is available both in Basic and Python user scripts.

Change-Id: Ifc9ce07b76e72fe580baf2c8103387cab56a6ef0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124230
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>