tdf#145363 Qt reparent modal dialogs on show
Simply said, one can't have two modal dialogs open at the same
parent in Qt. All modal windows must open as a stack, each
parented to the previous one. This is kind of logical.

Unexpectedly Qt totally breaks, if you open two modal dialogs on
the same window. This happens, because the existing paragraph
style dialog and the "sub" "list style" dialog are both children
of their Writer window.

I'm not sure the additionally introduced QWidget-based parent
handling is strictly needed. It seems Ok.

So for every visibility and modality change, we reparent the
Qt widget, either on top of the modal stack or restore the
original LO-based parent. The LO hierachy is never changed!

Change-Id: Id209c9aa67774440089dc50a6648ac293950087a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124500
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>