SwNavigator: Restore content tree settings before showing
Content tracking by the Navigator content tree and outline levels
displayed currently do not use the stored Navigator settings when the
tree is initially shown. This can cause content to be tracked and
outline levels to be shown that are different from the stored settings
when the tree is initially shown.

This patch fixes these possibilities by changing when the stored
content tree settings are applied to before the tree is initially
filled and shown.

Change-Id: I5443788c3fbacb36695cfc56163f217408254fd6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124831
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>