SwNavigator: Improve which entry is selected in the content tree
after the tree is updated and when current document cursor position is
at content that is set to not track

Intended behavior when the content tree display is updated is for the
selected entry before update to be reselected after update if the entry
is still in the tree. Currently, if the selected entry before update is
a content type, the selected entry after update is the first entry in
the tree. If the selected entry before update is not present, e.g. the
selected entry is the only member of a content type and is deleted, the
first entry in the tree is selected. In these cases, this patch changes
the selected entry after update to be the content type entry.

Instead of reselecting by entry name, which can result in incorrect
selection when non unique entry names are possible, as in Headings,
this patch improves entry reselect by comparing unique entry user data.

Content tracking is improved by not falling through to other tracked
content, e.g. when tables are not being tracked and sections are being
tracked, clicking in a table in a section will not select the section
in the content tree, the selected entry in the content tree will be
unchanged.

Change-Id: I8d82059db8688c31c7c0d705445c1d8b4272bec6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125176
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>