Use more precise value for M_PI in OpenCL
No idea if that would make actual difference, but it simply looks
odd that we use such a value, in view of all other efforts to improve
accuracy of rad<->deg conversions.

The value is used that is already present elsewhere in the code, e.g.
in sc/source/core/opencl/op_statistical.cxx.
Change-Id: I098f2bf748670e3250035d74d1bf7d1e224dd66b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125557
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>