tdf#123139 sc oox: CellXf should default to applyAlignment
Since the same logic applied to applyProtection
I tested and changed that one too.

No documentation was found to support the comment
(which was probably taken from binary import)
that applyAlignment had a context-dependent default value.

Testing on the other hand suggests that applyAlignment
is always considered true in the case of CellXf
and only matters for CellStyleXf.
The unit test tests for this - confirmed on Excel 2003/2016.

No existing unit tests where found with an
assert(rXf1.maModel.mbAlignUsed ||
(rXf1.maAlignment.getApiData() == rXf2.maAlignment.getApiData()));

Change-Id: I47d8dded93335092c93e75b4c18b798569da77f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126177
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>