tdf#129629 sw UI: Find/replace - stop popping open MORE section
This was a regression from LO 6.1, where it would always
show the "more" options if one of them was enabled.

That might be fine for the dialog initialization
(and that is how I implemented it)
but even that is a bit much. Perhaps it should be a
search option that is remembered while the application
is running. Is that what pSearchOptions holds?

Anyway, this much I can easily figure out.
Someone who knows more about these things
can enhance it further if they want.

Change-Id: I8288bea514840f4ea06deb00a3603aa9568a75b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126226
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>