WASM adapt link and debug flags
Changes the linker flags, so a debug build will not rewrite the
WASM binary code for instrumentation, as this takes ages and more
then 64GB of RAM, currently.

For the same reason - long link times - prefer DWARF debugging to
the external source maps.

Initial memory and thread pool size are linker flags, so move them.

Change-Id: I5cf39842e5091ef55b4fac719f881575e69fe8d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128589
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>